refactor: extract RichTextUndoState into standalone class to completely decouple from RichTextState
