+++ richtext-editor/src/main/kotlin/dev/mkeeda/arranger/richtext/editor/RichTextState.kt
@@ -230,8 +230,7 @@ public class RichTextState(initialText: RichString) {
             return
         }
 
-        val snapshotBefore = takeSnapshot()
-        val mergePolicy = resolveMergePolicy(buffer)
+        undoState.captureSnapshotBeforeChange(buffer)
 
         val typingAttr = typingAttributes
         val removedAttr = removedTypingAttributes
@@ -304,53 +303,15 @@ public class RichTextState(initialText: RichString) {
 
         this.spans = newSpans.resnapParagraphSpans(buffer.toString())
         clearTypingAttributes()
-
-        undoManager.pushSnapshot(snapshotBefore, mergePolicy)
     }
 
-    private fun takeSnapshot(): EditorSnapshot {
-        return EditorSnapshot(
-            text = textFieldState.text.toString(),
-            spans = spans,
-            selection = textFieldState.selection,
+    public val undoState: RichTextUndoState =
+        RichTextUndoState(
+            textFieldState = textFieldState,
+            getSpans = { spans },
+            setSpans = { spans = it },
+            clearTypingAttributes = { clearTypingAttributes() },
         )
-    }
-
-    @OptIn(ExperimentalFoundationApi::class)
-    private fun restoreSnapshot(snapshot: EditorSnapshot) {
-        textFieldState.edit {
-            replace(0, length, snapshot.text)
-            selection = snapshot.selection
-        }
-        textFieldState.undoState.clearHistory()
-        this.spans = snapshot.spans
-        clearTypingAttributes()
-    }
-
-    public val undoState: RichTextUndoState = RichTextUndoState()
-
-    public inner class RichTextUndoState {
-        public val canUndo: Boolean get() = undoManager.canUndo
-        public val canRedo: Boolean get() = undoManager.canRedo
-
-        @OptIn(ExperimentalFoundationApi::class)
-        public fun undo() {
-            val undoneSnapshot = undoManager.undo(takeSnapshot()) ?: return
-            restoreSnapshot(undoneSnapshot)
-        }
-
-        @OptIn(ExperimentalFoundationApi::class)
-        public fun redo() {
-            val redoneSnapshot = undoManager.redo(takeSnapshot()) ?: return
-            restoreSnapshot(redoneSnapshot)
-        }
-
-        @OptIn(ExperimentalFoundationApi::class)
-        public fun clearHistory() {
-            undoManager.clear()
-            textFieldState.undoState.clearHistory()
-        }
-    }
 
     private fun handleNewlineInsertion(
         insertedText: String,
@@ -559,26 +520,4 @@ private fun shiftSpan(
             }
         }
     }
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-internal fun resolveMergePolicy(buffer: TextFieldBuffer): UndoMergePolicy {
-    if (buffer.changes.changeCount != 1) return UndoMergePolicy.Separate
-
-    val range = buffer.changes.getRange(0)
-    val originalRange = buffer.changes.getOriginalRange(0)
-
-    val insertedText = buffer.asCharSequence().substring(range.min, range.max)
-    val deletedLength = originalRange.length
-
-    // Deletion always creates a new undo entry
-    if (deletedLength > 0) return UndoMergePolicy.Separate
-
-    // Paste (multiple characters) always creates a new undo entry
-    if (insertedText.length > 1) return UndoMergePolicy.Separate
-
-    // Space or newline creates a new undo entry
-    if (insertedText.any { it.isWhitespace() }) return UndoMergePolicy.Separate
-
-    return UndoMergePolicy.Merge
 }
