+++ richtext-editor/src/main/kotlin/dev/mkeeda/arranger/richtext/editor/RichTextUndoState.kt
@@ -1,0 +1,89 @@
+package dev.mkeeda.arranger.richtext.editor
+
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.text.input.TextFieldBuffer
+import androidx.compose.foundation.text.input.TextFieldState
+import androidx.compose.runtime.Stable
+import dev.mkeeda.arranger.richtext.RichSpan
+
+@Stable
+public class RichTextUndoState internal constructor(
+    private val textFieldState: TextFieldState,
+    private val getSpans: () -> List<RichSpan>,
+    private val setSpans: (List<RichSpan>) -> Unit,
+    private val clearTypingAttributes: () -> Unit,
+) {
+    internal val undoManager = RichTextUndoManager()
+
+    public val canUndo: Boolean
+        get() = undoManager.canUndo
+
+    public val canRedo: Boolean
+        get() = undoManager.canRedo
+
+    @OptIn(ExperimentalFoundationApi::class)
+    public fun undo() {
+        val undoneSnapshot = undoManager.undo(takeSnapshot()) ?: return
+        restoreSnapshot(undoneSnapshot)
+    }
+
+    @OptIn(ExperimentalFoundationApi::class)
+    public fun redo() {
+        val redoneSnapshot = undoManager.redo(takeSnapshot()) ?: return
+        restoreSnapshot(redoneSnapshot)
+    }
+
+    @OptIn(ExperimentalFoundationApi::class)
+    public fun clearHistory() {
+        undoManager.clear()
+        textFieldState.undoState.clearHistory()
+    }
+
+    @OptIn(ExperimentalFoundationApi::class)
+    internal fun captureSnapshotBeforeChange(buffer: TextFieldBuffer) {
+        val snapshotBefore = takeSnapshot()
+        val mergePolicy = resolveMergePolicy(buffer)
+        undoManager.pushSnapshot(snapshotBefore, mergePolicy)
+    }
+
+    private fun takeSnapshot(): EditorSnapshot {
+        return EditorSnapshot(
+            text = textFieldState.text.toString(),
+            spans = getSpans(),
+            selection = textFieldState.selection,
+        )
+    }
+
+    @OptIn(ExperimentalFoundationApi::class)
+    private fun restoreSnapshot(snapshot: EditorSnapshot) {
+        textFieldState.edit {
+            replace(0, length, snapshot.text)
+            selection = snapshot.selection
+        }
+        textFieldState.undoState.clearHistory()
+        setSpans(snapshot.spans)
+        clearTypingAttributes()
+    }
+}
+
+@OptIn(ExperimentalFoundationApi::class)
+internal fun resolveMergePolicy(buffer: TextFieldBuffer): UndoMergePolicy {
+    if (buffer.changes.changeCount != 1) return UndoMergePolicy.Separate
+
+    val range = buffer.changes.getRange(0)
+    val originalRange = buffer.changes.getOriginalRange(0)
+
+    val insertedText = buffer.asCharSequence().substring(range.min, range.max)
+    val deletedLength = originalRange.length
+
+    // Deletion always creates a new undo entry
+    if (deletedLength > 0) return UndoMergePolicy.Separate
+
+    // Paste (multiple characters) always creates a new undo entry
+    if (insertedText.length > 1) return UndoMergePolicy.Separate
+
+    // Space or newline creates a new undo entry
+    if (insertedText.any { it.isWhitespace() }) return UndoMergePolicy.Separate
+
+    return UndoMergePolicy.Merge
+}
